requirements.txt
Django==5.1
mysql-connector-python == 9.0.0
mysqlclient == 2.2.4
django-rest-framework
openai
python-dotenv
django-cors-headers == 4.4.0
pyJWT
django-storages
azure-storage-file-share
azure-identity
django-two-factor-auth
django-phonenumber-field[phonenumbers]
django-crispy-forms
twilio
django-auto-logout
pytz